backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
apscheduler==3.10.4
minio==7.2.12
openai==1.58.1
beautifulsoup4==4.12.3
imagehash==4.3.2
Pillow==11.1.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
python-multipart==0.0.20
cryptography==44.0.0
curl_cffi==0.7.4
